# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "代理订阅售卖系统后端 API。统一信封 {code,message,data}；鉴权 Bearer \u003caccess_token\u003e。",
        "title": "NanoCloud API",
        "contact": {},
        "version": "1.0"
    },
    "host": "api.example.com",
    "basePath": "/api/v1",
    "paths": {
...
        "/user/password/change": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ChangePasswordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/resp.Body"
                        }
                    }
                }
            }
        },
        "/user/profile": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "更新通知设置",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateProfileReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.UserProfileResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/stat": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户仪表板统计",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.UserStatResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.ChangePasswordReq": {
            "type": "object",
            "required": [
                "new_password",
                "old_password"
            ],
            "properties": {
                "new_password": {
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 8
                },
                "old_password": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateProfileReq": {
            "type": "object",
            "properties": {
                "remind_expire": {
                    "type": "boolean"
                },
                "remind_traffic": {
                    "type": "boolean"
                }
            }
        },
...
        "model.UserProfileResp": {
            "type": "object",
            "properties": {
                "remind_expire": {
                    "type": "boolean"
                },
                "remind_traffic": {
                    "type": "boolean"
                }
            }
        },
        "model.UserStatResp": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "commission_balance": {
                    "type": "number"
                },
                "email": {
                    "type": "string"
                },
                "invited_count": {
                    "type": "integer"
                },
                "is_agent": {
                    "type": "boolean"
                },
                "open_ticket_count": {
                    "type": "integer"
                },
                "pending_order_count": {
                    "type": "integer"
                }
            }
        },
        "resp.Body": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}